% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseFlowIndex.R
\name{BaseFlowIndex}
\alias{BaseFlowIndex}
\title{Calculates Base Flow Index}
\usage{
BaseFlowIndex(Qflow, timestep = 1/24)
}
\arguments{
\item{Qflow}{time series of streamflow discharge}

\item{timestep}{time step for the measured data (equally spaced)}
}
\value{
Curve Number, in the range [0,100]
}
\description{
Calculates Base Flow Index
}
\details{
Calculates Base Flow Index for a given discharge time series as described in Report # 108, Low flow estimation in the United Kingdom, by A. Gustard, A. Bullock, and J.M. Dickson, 1992
}
\examples{
# BaseFlowIndex(Q)
}
\author{
Nataliya Bulygina (July, 2008)
}

