% Generated by roxygen2 (4.0.2): do not edit by hand
\name{CalculateQ}
\alias{CalculateQ}
\title{Calculate the direct storm runoff (Q) from P and CN}
\usage{
CalculateQ(P, CN, PQunits = "mm")
}
\arguments{
\item{P}{precipitation}

\item{CN}{curve number, in the range [0,100]}

\item{PQunits}{units in which P and Q are expressed (default="mm")}
}
\value{
Q, direct storm runoff.
}
\description{
Calculate the direct storm runoff (Q) from P and CN
}

