% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/centroid.R
\name{centroid}
\alias{centroid}
\title{Calculate centroid of time series events}
\usage{
centroid(dataX, infoX, useMax = FALSE, altStart = NULL)
}
\arguments{
\item{dataX}{is a time series}

\item{infoX}{table containing summary of events for the time series dataX}

\item{useMax}{boolean, if FALSE (default) the centroid of the event is calculated. If TRUE, the centroid is assumed to correspond to the following maximum. This is generally used with Q time series.}

\item{altStart}{vector containing the timestamp of alternative starting dates for the events. This is generally used with Q time series and set to the P centroid.}
}
\value{
Updated infoX table containing the time and index of centroid.
}
\description{
Calculate centroid of time series events
}
\examples{
# newInfoP <- centroid(dataP, infoP)
# newInfoQ <- centroid(data$Q, infoQ, useMax = TRUE, altStart = newInfoP$indexCentroid)
}

