% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClipSoilMap.R
\name{ClipSoilMap}
\alias{ClipSoilMap}
\title{Clip soil map over a polygon and (re)calculate areas.}
\usage{
ClipSoilMap(soil, catchment)
}
\arguments{
\item{soil}{SpatialPolygonsDataFrame containing percentage of HOST soil classes, i.e. see data("PlynlimonSOIL").}

\item{catchment}{SpatialPolygonsDataFrame containing a single catchment boundary.}
}
\value{
A SpatialPolygonsDataFrame clipped over a catchment with an additional column with updated areas.
}
\description{
Clip soil map over a polygon and (re)calculate areas.
}
\examples{
\dontrun{
  data("PlynlimonSUBCATCHMENTS")
  data("PlynlimonSOIL")
  soilMap <- ClipSoilMap(soil = PlynlimonSOIL,
                         catchment = PlynlimonSUBCATCHMENTS[1,])
}

}

