% Generated by roxygen2 (4.0.2): do not edit by hand
\name{CNfromMaps}
\alias{CNfromMaps}
\title{Calculate the Curve Number from HOST soil map}
\usage{
CNfromMaps(soilMap, tabulatedCN, shpFolder = NULL, mask = NULL,
  plotOption = FALSE)
}
\arguments{
\item{soilMap}{filename of a single band raster map (e.g. tif) that contains HOST soil classes (these are integers in the range [1,29]). By default this should be in the working directory.}

\item{tabulatedCN}{list containing the CN tabulated values for the four hydrologic soil group (e.g. list("A"=30,"B"=55,"C"=70,"D"=77)). This is tabulated here: http://en.wikipedia.org/wiki/Runoff_curve_number.}

\item{shpFolder}{(optional) the path to the folder containing the GIS layers. By default this is the working directory.}

\item{mask}{(optional) filename of a shapefile containing the a single polygon to be used as mask to crop the raster. This is usually the catchment boundary.}

\item{plotOption}{if TRUE it plots the soil map and mask}
}
\value{
Curve Number, in the range [0,100]
}
\description{
Calculate the Curve Number from HOST soil map
}
\examples{
# CNfromMaps(soilMap="PlynlimonSoils.tif",
# tabulatedCN=list("A"=30,"B"=55,"C"=70,"D"=77),
# shpFolder="/home/claudia/Dropbox/Projects/PURE/PURE_shared/Data/",
# mask="severn", plotOption=FALSE)
}

