\name{curvenumber-package}
\alias{curvenumber-package}
\alias{curvenumber}
\docType{package}
\title{
SCS Curve Number method
}
\description{
Implementation of the SCS curve number method.
}
\details{
\tabular{ll}{
Package: \tab curvenumber\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-11-09\cr
License: \tab What license is it under?\cr
}
Implementation of the SCS curve number method.
}
\author{
Claudia Vitolo
Maintainer: Claudia Vitolo <cvitolodev@gmail.com>
}
\references{
R.H. Hawkins (1993), 'Asymptotic determination of runoff curve numbers from data', J. Irrig. Drain.Eng. ASCE 119 (2), 334-345.

}
\keyword{ package }
