\name{DATA}
\alias{DATA}
\docType{data}
\title{
DATA
}
\description{
sample dataset to test the functions in the curvenumber package
}
\usage{data("DATA")}
\format{
  The format is: zoo series from 2007-09-28 10:00:00 to 2009-12-01 11:00:00. made of 3 columns: "P" "E" "Q"
}
\examples{
data(DATA)
## maybe str(DATA) ; plot(DATA) ...
}
\keyword{datasets}
