% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectStormRunoff.R
\name{DirectStormRunoff}
\alias{DirectStormRunoff}
\title{Calculate the direct storm runoff (Q) from P and CN}
\usage{
DirectStormRunoff(P, CN, PQunits = "mm")
}
\arguments{
\item{P}{precipitation}

\item{CN}{curve number, in the range [0,100]}

\item{PQunits}{units in which P and Q are expressed (default="mm")}
}
\value{
list containing two elements: S () and Q (direct storm runoff).
}
\description{
Calculate the direct storm runoff (Q) from P and CN
}
\examples{
\dontrun{
  SQ <- DirectStormRunoff(P = 30, CN = 80, PQunits = "mm")
}

}

