% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmpiricalCN.R
\name{EmpiricalCN}
\alias{EmpiricalCN}
\title{Calculate the Curve Number from time series data (P and Q)}
\usage{
EmpiricalCN(tseries, stepsBack = 5, hours2extend = 6, timeUnits = "hours",
  verbose = FALSE)
}
\arguments{
\item{tseries}{time series object containing, at least, precipitation ("P") and discharge ("Q")}

\item{stepsBack}{timesteps to use to generate the linear model for the volume under the rising limb (default = 5)}

\item{hours2extend}{number of hours to extend the rainfall event to account for the discharge peak to occur (default = 6)}

\item{timeUnits}{time step units (default = "hours")}

\item{verbose}{(optional) boolean (FALSE by default). If TRUE prints the progress in terms of event number.}
}
\value{
Empirical Curve Number for a given catchment, this is an integer in the range [0,100] with 0 = no runoff and 100 = all rainfall becomes runoff.
}
\description{
Calculate the Curve Number from time series data (P and Q)
}
\examples{
\dontrun{
  data("SevernTS")
  severnCN <- EmpiricalCN(tseries = SevernTS)
}

}

