% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowSeparation.R
\name{FlowSeparation}
\alias{FlowSeparation}
\title{Separate Baseflow and Surface runoff}
\usage{
FlowSeparation(DATA, eventTable, stepsBack = 5, timeUnits = "hours",
  plotOption = FALSE, event2plot = 1, verbose = FALSE)
}
\arguments{
\item{DATA}{data.frame containing P, Q (and E)}

\item{eventTable}{table containing summary of Q events}

\item{stepsBack}{timesteps to use to generate the linear model for the volume under the rising limb (default = 5)}

\item{timeUnits}{time step units (default = "hours")}

\item{plotOption}{boolean, if TRUE (default) it prints a plot to show the division of the events}

\item{event2plot}{even number for which the plot should be generated (default = 3). This is only used if plotOption = TRUE}

\item{verbose}{(optional) boolean (FALSE by default). If TRUE prints the progress in terms of event number.}
}
\value{
updated table containing summary of Q events with 4 more columns:
volumeQ, baseflowVolume, surfaceVolume and surfacePeak
}
\description{
Separate Baseflow and Surface runoff
}
\details{
From Boorman 1995: The recession prior to the event is continued through the event, and this flow is subtracted from the total flow hydrograph. A straight line is then drawn from beneath the peak flow, or centroid of peaks, to the point already identified as marking the end of response runoff. The response runoff is the portion of flow above this separation.
}
\examples{
\dontrun{
  data("SevernTS")
  tableP <- FindPevents(SevernTS$P[1:1000])
  tableQ <- FindQevents(SevernTS$Q[1:1000], tableP, hours2extend=6)
  df <- FlowSeparation(SevernTS, tableQ, stepsBack=5, timeUnits = "hours",
                       plotOption = TRUE, event2plot = 3)
}

}

