% Generated by roxygen2 (4.0.2): do not edit by hand
\name{HOST2USDA}
\alias{HOST2USDA}
\title{Map HOST classes into USDA classes}
\usage{
HOST2USDA(vectorHOSTclass)
}
\arguments{
\item{vectorHOSTclass}{this is a vector of HOST classes (e.g. c("3","7")). This can contain a single value.}
}
\value{
USDA class, this can be: A, B, C, D or a combination of them.
}
\description{
Map HOST classes into USDA classes
}
\examples{
HOST2USDA(c("3","7"))
}

