% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HOSTband2class.R
\name{HOSTband2class}
\alias{HOSTband2class}
\title{Map raster band values into HOST classes}
\usage{
HOSTband2class(rasterBandValue)
}
\arguments{
\item{rasterBandValue}{this is a vector of raster band values (e.g. c("3","7")). This can contain a single value. Each value is an integer in the range [1,12].}
}
\value{
HOST class, this can be an integer (or a vector of inegers) in the range [1,29].
}
\description{
Map raster band values into HOST classes
}
\examples{
# HOSTband2class(c("3","7"))

}
