% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lagtime.R
\name{lagtime}
\alias{lagtime}
\title{Calculate lagtime between precipitation and discharge time series}
\usage{
lagtime(newInfoP, newInfoQ, timeUnits = "hours")
}
\arguments{
\item{newInfoP}{table containing summary of P events}

\item{newInfoQ}{table containing summary of Q events}

\item{timeUnits}{string defining the time unit to express the lagtime (default = "hours")}
}
\value{
lagtime in the units specified by timeUnits.
}
\description{
Calculate lagtime between precipitation and discharge time series
}
\examples{
# LagPQ <- lagtime(tableP, tableQ, timeUnits = "hours")
}

