% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeLoopkupTable.R
\name{MakeLoopkupTable}
\alias{MakeLoopkupTable}
\title{Generate a lookup table for soil, vegetation and hydrological properties}
\usage{
MakeLoopkupTable(case = "default")
}
\arguments{
\item{case}{string of characters. It can be one of the following options: "default", "Severn&Wye", "Good woods + Good pasture", "Good woods + Fair pasture", "Fair woods + Poor pasture".}
}
\value{
data.frame containing 10 rows (vegetation classes) and the following columns: Class (vegetation), Subclass (vegetation), Code, SoilA, SoilB, SoilC, SoilD.
}
\description{
Generate a lookup table for soil, vegetation and hydrological properties
}
\examples{
\dontrun{
  dfLookup <- MakeLoopkupTable()
  dfLookupSW <- MakeLoopkupTable("Severn&Wye")
  dfLookupGWGP <- MakeLoopkupTable("Good woods + Good pasture")
  dfLookupGWFP <- MakeLoopkupTable("Good woods + Fair pasture")
  dfLookupFWPP <- MakeLoopkupTable("Fair woods + Poor pasture")
}

}
