% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PercentageCoverage.R
\name{PercentageCoverage}
\alias{PercentageCoverage}
\title{Calculate the Curve Number from HOST soil map}
\usage{
PercentageCoverage(myRaster, shpFolder = NULL, myMask = NULL,
  myCRS = NULL, host = FALSE, methodBy = "Bulygina et al., 2011",
  plotOption = FALSE)
}
\arguments{
\item{myRaster}{filename of a single band raster map (e.g. tif) that contains classes. By default this should be in the working directory.}

\item{shpFolder}{(optional) the path to the folder containing the GIS layers. By default this is the working directory.}

\item{myMask}{(optional) filename of a shapefile containing the a single polygon to be used as mask to crop the raster. This is usually the catchment boundary.}

\item{myCRS}{proj4 for the desired Coordinate Reference System (default = British National Grid). See http://spatialreference.org/}

\item{host}{if TRUE it uses HOST classes, otherwise it assumes that the classification is based on the Soil Hydrology Study conducted by JP Bell in 1968-1969: Bell, J.P. (1969). The Soil Hydrology of the Plynlimon Catchments. Institute of Hydrology Report No. 8, Institute of Hydrology, Wallingford, UK.}

\item{methodBy}{this is a string that identifies the methodology to be used. Possible values are: "Bulygina et al., 2011" (default), "Halcrow, 2011".}

\item{plotOption}{if TRUE it plots the soil map and mask}
}
\value{
table containing the percentage coverages
}
\description{
Calculate the Curve Number from HOST soil map
}
\examples{
# PercentageCoverage(myRaster=mySoilMap1000,shpFolder=myShpFolder,myMask="severn")

}

