% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PercentageCoverageSoilClasses.R
\name{PercentageCoverageSoilClasses}
\alias{PercentageCoverageSoilClasses}
\title{Percentage coverage of soil classes.}
\usage{
PercentageCoverageSoilClasses(soilMap, ignoreColumns = NULL,
  removeZeros = TRUE)
}
\arguments{
\item{soilMap}{SpatialPolygonsDataFrame containing percentage of HOST soil classes, i.e. see data("PlynlimonSOIL").}

\item{ignoreColumns}{(optional) vector containing the columns to ignore}

\item{removeZeros}{(optional) boolean. If TRUE (default), the function returns only non zero classes.}
}
\value{
A named numeric vector containing the percentage coverage of soil classes.
}
\description{
Percentage coverage of soil classes.
}
\details{
This function recognises the column containing soil classes (it assumes that the name of the column contains the string "code").
}
\examples{
\dontrun{
  data("PlynlimonSUBCATCHMENTS")
  data("PlynlimonSOIL")
  # Clip soil map over Severn catchment
  soilMap <- ClipSoilMap(soil = PlynlimonSOIL,
                         catchment = PlynlimonSUBCATCHMENTS[1,])
  PercentageCoverageSoilClasses(soilMap=soilMap, ignoreColumns = c(32, 33))
}

}

