% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotEvents.R
\name{PlotEvents}
\alias{PlotEvents}
\title{Plot PQ events}
\usage{
PlotEvents(tseries, eventTable)
}
\arguments{
\item{tseries}{time series object containing, at least, precipitation ("P") and discharge ("Q")}

\item{eventTable}{table containing summary of P&Q events}
}
\value{
Plot showing PQ events
}
\description{
Plot PQ events
}
\examples{
\dontrun{
  data("SevernTS")
  tableP <- FindPevents(SevernTS$P[1:100])
  tableQ <- FindQevents(Q = SevernTS$Q[1:100],
                        eventTableP = tableP,
                        hours2extend = 6)

  PlotEvents(tseries = SevernTS[1:100,],
             eventTable = tableQ)
}

}

