% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvenumber-package.R
\docType{data}
\name{PlynlimonSOILdominant}
\alias{PlynlimonSOILdominant}
\title{Data set: PlynlimonSOILdominant}
\format{Formal class 'RasterLayer' [package "raster"] made of a single layer. Values range between 0 and 29. For correspondence to USDA classes see S1 lookup table.}
\source{
\url{http://hdl.handle.net/10044/1/30773}
}
\usage{
data("PlynlimonSOILdominant")
}
\description{
PlynlimonSOILdominant is a raster layer containing the dominant soil classes for Plynlimon catchments (based on the HOST classification).
}
\references{
Vitolo, C. (2015), Exploring data mining for hydrological modelling (Ph.D. thesis), Imperial College London, uri: http://hdl.handle.net/10044/1/30773.
}
\keyword{datasets}

