% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvenumber-package.R
\docType{data}
\name{PlynlimonTable}
\alias{PlynlimonTable}
\title{Data set: PlynlimonTable}
\format{
A data frame with 9 rows and the following 3 columns:
\describe{
  \item{\code{ID}}{Subcatchment identification code (character)}
  \item{\code{Name}}{Full name of subcatchment (character)}
  \item{\code{ShortName}}{Short name of subcatchment (character)}
}
}
\source{
\url{http://hdl.handle.net/10044/1/30773}
}
\usage{
data("PlynlimonTable")
}
\description{
PlynlimonTable is a look-up table for Plynlimon subcatchment names and id codes.
}
\references{
Vitolo, C. (2015), Exploring data mining for hydrological modelling (Ph.D. thesis), Imperial College London, uri: http://hdl.handle.net/10044/1/30773.
}
\keyword{datasets}
