% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvenumber-package.R
\docType{data}
\name{PlynlimonVEG}
\alias{PlynlimonVEG}
\title{Data set: PlynlimonVEG}
\format{
Formal class 'RasterLayer' [package "raster"] made of a single layer. Values range between 0 and 10. For correspondence to vegetation classes see \code{PlynlimonVegSoilTable} look up table.
}
\source{
\url{http://hdl.handle.net/10044/1/30773}
}
\usage{
data("PlynlimonVEG")
}
\description{
PlynlimonVEG is a vegetation map for Plynlimon are. It is extracted from the Vegetation map 2013 available from CEH data portal.
}
\references{
Vitolo, C. (2015), Exploring data mining for hydrological modelling (Ph.D. thesis), Imperial College London, uri: http://hdl.handle.net/10044/1/30773.
}
\keyword{datasets}
