% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvenumber-package.R
\docType{data}
\name{PlynlimonVegSoilTable}
\alias{PlynlimonVegSoilTable}
\title{Data set: PlynlimonVegSoilTable}
\format{A data frame with 58 rows and the following 12 columns:
\describe{
  \item{\code{Class}}{Description of vegetation class}
  \item{\code{SubClass}}{Description of vegetation subclass}
  \item{\code{LCM2007Class}}{Corresponding class in Land Cover Map 2007}
  \item{\code{LCM2007SubClass}}{Corresponding subclass in Land Cover Map 2007}
  \item{\code{LCM2000Class}}{Corresponding class in Land Cover Map 2000}
  \item{\code{LCM2000SubClass}}{Corresponding subclass in Land Cover Map 2000}
  \item{\code{LCM1990SubClass}}{Corresponding subclass in Land Cover Map 1990}
  \item{\code{VEG2013SubClass}}{Corresponding subclass in Vegetation Map 2013}
  \item{\code{SoilA}}{CN value for soil type A}
  \item{\code{SoilB}}{CN value for soil type B}
  \item{\code{SoilC}}{CN value for soil type C}
  \item{\code{SoilD}}{CN value for soil type D}
}}
\source{
\url{http://hdl.handle.net/10044/1/30773}
}
\usage{
data("PlynlimonVegSoilTable")
}
\description{
PlynlimonVegSoilTable is a look-up table for vegetation and soil classes.
}
\references{
Vitolo, C. (2015), Exploring data mining for hydrological modelling (Ph.D. thesis), Imperial College London, uri: http://hdl.handle.net/10044/1/30773.
}
\keyword{datasets}

