% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionalisedBFI.R
\name{RegionalisedBFI}
\alias{RegionalisedBFI}
\title{Calculate regionalised Base Flow Index (BFI) from percentages of soil classes.}
\usage{
RegionalisedBFI(soil, catchment, lookupTable)
}
\arguments{
\item{soil}{SpatialPolygonsDataFrame containing percentage of soil classes, this can be the HOST classification for the UK, see data("PlynlimonSOIL").}

\item{catchment}{SpatialPolygonsDataFrame containing a single catchment boundary.}

\item{lookupTable}{A dataframe containing as many rows as the number of soil classes and at least one column (named "BFIHOST").}
}
\value{
A numeric value representing the regionalised BFI.
}
\description{
Calculate regionalised Base Flow Index (BFI) from percentages of soil classes.
}
\examples{
\dontrun{
  # Load Plynlimon sub-catchments spatial polygons data frame
  data("PlynlimonSUBCATCHMENTS")
  # Load soil map (percentage distribution of classes)
  data("PlynlimonSOIL")
  # Load lookup table
  data("S1")

  RegionalisedBFI(soil = PlynlimonSOIL,
                  catchment = PlynlimonSUBCATCHMENTS[1,],
                  lookupTable = S1)
}

}
