% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionalisedBFIdominant.R
\name{RegionalisedBFIdominant}
\alias{RegionalisedBFIdominant}
\title{Calculate Base Flow Index (BFI) map from soil (HOST) map}
\usage{
RegionalisedBFIdominant(soil, catchment, lookupTable)
}
\arguments{
\item{soil}{filename of a raster containing the dominant soil classes (HOST).}

\item{catchment}{SpatialPolygonsDataFrame containing a single catchment boundary to be used as mask to crop the raster.}

\item{lookupTable}{A dataframe containing as many rows as the number of soil classes and at least two columns (named "HOSTclass" and "BFIHOST").}
}
\value{
Regionalised BFI number for a given catchment, this is in the range [0,1] and is calculated only using dominant soil classes.
}
\description{
Calculate Base Flow Index (BFI) map from soil (HOST) map
}
\examples{
\dontrun{
  # Load Plynlimon sub-catchments spatial polygons data frame
  data("PlynlimonSUBCATCHMENTS")
  # Load soil map (HOST dominant classes)
  data("PlynlimonSOILdominant")
  # Load lookup table
  data("S1")

  bfi <- RegionalisedBFIdominant(soil = PlynlimonSOILdominant,
                         catchment = PlynlimonSUBCATCHMENTS[1,],
                         lookupTable = S1)
}

}

