% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionalisedCN.R
\name{RegionalisedCN}
\alias{RegionalisedCN}
\title{Calculate Curve Number for soil (HOST) and vegetation (LCM) maps}
\usage{
RegionalisedCN(soil, catchment, lookupTable, vegetation,
  artificialDrainage = "none")
}
\arguments{
\item{soil}{filename of a shapefile containing the percentage coverage of HOST classes}

\item{catchment}{SpatialPolygonsDataFrame containing a single catchment boundary to be used as mask to crop the raster.}

\item{lookupTable}{this is a data.frame to link soil, vegetation and Hydrological conditions.}

\item{vegetation}{filename of a raster containing the vegetation map.}

\item{artificialDrainage}{(default = "none"), possible values are none, low, medium, high.}
}
\value{
Regionalised Curve Number for a given catchment, this is an integer in the range [0,100] with 0 = no runoff and 100 = all rainfall becomes runoff.
}
\description{
Calculate Curve Number for soil (HOST) and vegetation (LCM) maps
}
\details{
Please make sure that soil, catchment and vegetation are projected in the same Coordinate Reference System (for info see http://spatialreference.org/).
}
\examples{
\dontrun{
  # Load Plynlimon sub-catchments spatial polygons data frame
  data("PlynlimonSUBCATCHMENTS")
  # Load soil map (HOST percentage distribution of classes)
  data("PlynlimonSOIL")
  # Load land cover (VEGETATION 2013) MAP
  data("PlynlimonVEG")
  # Load lookup table
  dfLookup <- MakeLoopkupTable("Severn&Wye")

  cn <- RegionalisedCN(soil = PlynlimonSOIL,
                       catchment = PlynlimonSUBCATCHMENTS[1,],
                       lookupTable = dfLookup,
                       vegetation = PlynlimonVEG,
                       artificialDrainage = "none")
}

}

