% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionalisedCNdominant.R
\name{RegionalisedCNdominant}
\alias{RegionalisedCNdominant}
\title{Calculate Curve Number for soil dominant classes (HOST) and vegetation (LCM) maps}
\usage{
RegionalisedCNdominant(soil, catchment, lookupTable, vegetation,
  artificialDrainage = "none")
}
\arguments{
\item{soil}{filename of a raster containing the dominant soil classes (HOST).}

\item{catchment}{SpatialPolygonsDataFrame containing a single catchment boundary to be used as mask to crop the raster.}

\item{lookupTable}{this is a data.frame to link soil, vegetation and Hydrological conditions.}

\item{vegetation}{filename of a raster containing the vegetation map.}

\item{artificialDrainage}{(default = "none"), possible values are none, low, medium, high.}
}
\value{
Curve Number map, a GeoTiff raster with same resolution of soil and vegetation maps and extent equal to the mask
}
\description{
Calculate Curve Number for soil dominant classes (HOST) and vegetation (LCM) maps
}
\examples{
\dontrun{
  # Load Plynlimon sub-catchments spatial polygons data frame
  data("PlynlimonSUBCATCHMENTS")
  # Load soil map (HOST dominant classes)
  data("PlynlimonSOILdominant")
  # Load land cover (VEGETATION 2013) MAP
  data("PlynlimonVEG")
  # Load lookup table
  dfLookup <- MakeLoopkupTable("Fair woods + Poor pasture")

  cn <- RegionalisedCNdominant(soil = PlynlimonSOILdominant,
                               catchment = PlynlimonSUBCATCHMENTS[1,],
                               lookupTable = dfLookup,
                               vegetation = PlynlimonVEG,
                               artificialDrainage = "none")
}

}

