% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReturnPeriod.R
\name{ReturnPeriod}
\alias{ReturnPeriod}
\title{Calculate return periods for PQ events}
\usage{
ReturnPeriod(eventTable)
}
\arguments{
\item{eventTable}{is a table containing summary of P and Q events}
}
\value{
data.frame containing 3 columns: Tr (return period), P (max precipitation) and Q (max discharge)
}
\description{
Calculate return periods for PQ events
}
\examples{
# ReturnPeriod(eventTable)

}

