% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvenumber-package.R
\docType{data}
\name{SevernTS}
\alias{SevernTS}
\title{Data set: SevernTS}
\format{The time series object (e.g. ID54022) is of class zoo. The index is a POSIXct in the format "1975-04-27 03:00:00". The data is measured in mm/hour and contain the following variables:
\describe{
  \item{\code{P}}{Precipitation.}
  \item{\code{Q}}{Streamflow discharge}
}}
\source{
\url{http://hdl.handle.net/10044/1/30773}
}
\usage{
data("SevernTS")
}
\description{
SevernTS contains time series information for one of the gauged sub-catchment in the Plynlimon area: Severn at Plynlimon flume. The ID code of this subcatchment is 54022, see the look-up table \code{PlynlimonTable}.
}
\references{
Vitolo, C. (2015), Exploring data mining for hydrological modelling (Ph.D. thesis), Imperial College London, uri: http://hdl.handle.net/10044/1/30773.
}
\keyword{datasets}

