% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilandVeg2CN.R
\name{SoilandVeg2CN}
\alias{SoilandVeg2CN}
\title{Soil and Vegetation code to CN, and corrects it based on drainage level, if needed.}
\usage{
SoilandVeg2CN(soilCODE, vegetationCODE, lookupTable, artificialDrainage)
}
\arguments{
\item{soilCODE}{HOST soil class (integer in the range [1,29])}

\item{vegetationCODE}{integer in the range [1,10]}

\item{lookupTable}{this is a data.frame to link Soil, Vegetation and Hydrological conditions.}

\item{artificialDrainage}{(default = "none"), possible values are none, low, medium,high.}
}
\value{
CN, integer in the range [1,100]
}
\description{
Soil and Vegetation code to CN, and corrects it based on drainage level, if needed.
}
\examples{
# SoilandVeg2CN(15,6,lookupTable,artificialDrainage="none")

}

