% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StDevBFIHOST.R
\name{StDevBFIHOST}
\alias{StDevBFIHOST}
\title{Generate Base Flow Index (BFI) map from percentages of HOST classes.}
\usage{
StDevBFIHOST(soil, catchment, lookupTable)
}
\arguments{
\item{soil}{SpatialPolygonsDataFrame containing percentage of soil classes, this can be the HOST classification for the UK, see data("PlynlimonSOIL").}

\item{catchment}{SpatialPolygonsDataFrame containing a single catchment boundary.}

\item{lookupTable}{A dataframe containing as many rows as the number of soil classes and at least one column (named "BFIHOST").}
}
\value{
standard deviation of BFI.
}
\description{
Generate Base Flow Index (BFI) map from percentages of HOST classes.
}
\examples{
\dontrun{
  data("PlynlimonSUBCATCHMENTS")
  data("PlynlimonSOIL")
  data("S1")

  stdevBFI <- StDevBFIHOST(soil = PlynlimonSOIL,
                           catchment = PlynlimonSUBCATCHMENTS[1,],
                           lookupTable = S1)
}

}

