% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BFIHOST.R
\name{BFIHOST}
\alias{BFIHOST}
\title{Generate Base Flow Index (BFI) map from percentages of HOST classes.}
\usage{
BFIHOST(soilMap)
}
\arguments{
\item{soilMap}{shapefile that contains percentage of HOST soil classes.}
}
\value{
BFIHOST a number in the range [0,1] that represents the theoretical BFI.
}
\description{
Generate Base Flow Index (BFI) map from percentages of HOST classes.
}
\examples{
# BFI <- BFIHOST(soilMap=mySoilMap)
}

