% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CalculateCN.R
\name{CalculateCN}
\alias{CalculateCN}
\title{Calculate the Curve Number from P and Q}
\usage{
CalculateCN(dfTPQ, PQunits = "mm", plotOption = FALSE)
}
\arguments{
\item{dfTPQ}{data.frame containing 3 columns: Tr (return period), P (max precipitation) and Q (max discharge)}

\item{PQunits}{units in which P and Q are expressed (default="mm")}

\item{plotOption}{boolean, if TRUE (default) it prints a plot to show the division of the events}
}
\value{
Curve Number, in the range [0,100]
}
\description{
Calculate the Curve Number from P and Q
}

