% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNfromDATA.R
\name{CNfromDATA}
\alias{CNfromDATA}
\title{Calculate the Curve Number from time series data (P and Q)}
\usage{
CNfromDATA(DATA, stepsBack = 5, hours2extend = 6, timeUnits = "hours")
}
\arguments{
\item{DATA}{time series object containing, at least, precipitation ("P") and discharge ("Q")}

\item{stepsBack}{timesteps to use to generate the linear model for the volume under the rising limb (default = 5)}

\item{hours2extend}{number of hours to extend the rainfall event to account for the discharge peak to occur (default = 6)}

\item{timeUnits}{time step units (default = "hours")}
}
\value{
Curve Number, in the range [0,100]
}
\description{
Calculate the Curve Number from time series data (P and Q)
}
\examples{
# CNfromDATA(DATA)

}

