% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNfromMaps.R
\name{CNfromMaps}
\alias{CNfromMaps}
\title{Calculate Curve Number for soil (HOST) and vegetation (LCM) maps}
\usage{
CNfromMaps(Soil, Vegetation, Mask, lookupTable, myCRS = NULL,
  artificialDrainage = "none")
}
\arguments{
\item{Soil}{filename of a shapefile containing the percentage coverage of HOST classes}

\item{Vegetation}{filename of a raster containing the vegetation map.}

\item{Mask}{filename of a shapefile containing a single polygon to be used as mask to crop the raster. This is usually the catchment boundary.}

\item{lookupTable}{this is a data.frame to link Soil, Vegetation and Hydrological conditions.}

\item{myCRS}{proj4 for the desired Coordinate Reference System (default = British National Grid). See http://spatialreference.org/}

\item{artificialDrainage}{(default = "none"), possible values are none, low, medium,high.}
}
\value{
Curve Number map, a GeoTiff raster with same resolution of soil and vegetation maps and extent equal to the mask
}
\description{
Calculate Curve Number for soil (HOST) and vegetation (LCM) maps
}
\examples{
# CNmap <- CNfromMaps(Soil,Vegetation,Mask,lookupTable,
#                     myCRS=NULL,artificialDrainage="none",dominant=FALSE)

}

