% Generated by roxygen2 (4.0.2): do not edit by hand
\name{EventIdentification}
\alias{EventIdentification}
\title{Identify PQ events}
\usage{
EventIdentification(dataX, PQindependent = FALSE, plotOption = FALSE,
  variable2plot = "Q")
}
\arguments{
\item{dataX}{is a data.frame containing P, Q (and E) time series}

\item{PQindependent}{boolean. If TRUE the events P and Q are considered independent and the matching return periods are calculated using the frequency matching approach. If FALSE (default) the events P and Q are dependent and they have the same return period}

\item{plotOption}{boolean, if TRUE (default) it prints a plot to show the division of the events}

\item{variable2plot}{is a variable to plot}
}
\value{
table containing Return period and matching P and Q max values
}
\description{
Identify PQ events
}
\examples{
# df <- EventIdentification(dataX)
}

