% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/EventIdentification.R
\name{EventIdentification}
\alias{EventIdentification}
\title{Identify PQ events applying Frequency Matching approach}
\usage{
EventIdentification(dataX, hours2extend = 6, plotOption = FALSE,
  stepsBack = 5, timeUnits = "hours")
}
\arguments{
\item{dataX}{is a data.frame containing P, Q (and E) time series}

\item{hours2extend}{number of hours to extend the rainfall event to account for the discharge peak to occur}

\item{plotOption}{boolean, if TRUE (default) it prints a plot to show the division of the events}

\item{stepsBack}{timesteps to use to generate the linear model for the volume under the rising limb (default = 5)}

\item{timeUnits}{time step units (default = "hours")}
}
\value{
table containing Return period and matching P and Q max values
}
\description{
Identify PQ events applying Frequency Matching approach
}
\examples{
# df <- EventIdentification(dataX, hours2extend = 6, plotOption = FALSE)
}

