% Generated by roxygen2 (4.0.2): do not edit by hand
\name{findPevents}
\alias{findPevents}
\title{Find precipitation events}
\usage{
findPevents(dataX, plotOption = TRUE)
}
\arguments{
\item{dataX}{is a time series}

\item{plotOption}{boolean, if TRUE (default) it prints a plot to show the division of the events}
}
\value{
table containing summary of events for the time series dataX
}
\description{
Find precipitation events
}
\examples{
# tableP <- findPevents(data$P)
}

