% Generated by roxygen2 (4.0.2): do not edit by hand
\name{findPevents}
\alias{findPevents}
\title{Find precipitation events}
\usage{
findPevents(dataX)
}
\arguments{
\item{dataX}{is the table containing time series (P,Q and E)}
}
\value{
table containing summary of events for the time series dataX
}
\description{
Find precipitation events
}
\examples{
# tableP <- findPevents(dataX)
}

