% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPevents.R
\name{findPevents}
\alias{findPevents}
\title{Find precipitation events}
\usage{
findPevents(P)
}
\arguments{
\item{P}{is the precipitation time series}
}
\value{
table containing summary of events for the time series DATA
}
\description{
Find precipitation events
}
\examples{
# tableP <- findPevents(P)

}

