% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindPevents.R
\name{FindPevents}
\alias{FindPevents}
\title{Find precipitation events}
\usage{
FindPevents(P, verbose = FALSE)
}
\arguments{
\item{P}{is the precipitation time series}

\item{verbose}{(optional) boolean (FALSE by default). If TRUE prints the progress in terms of event number.}
}
\value{
table containing summary of events for the time series DATA
}
\description{
Find precipitation events
}
\examples{
\dontrun{
  data("SevernTS")
  tableP <- FindPevents(SevernTS$P)
}

}

