% Generated by roxygen2 (4.0.2): do not edit by hand
\name{findQevents}
\alias{findQevents}
\title{Find discharge events}
\usage{
findQevents(dataX, infoP, hours2extend = 8)
}
\arguments{
\item{dataX}{is the table containing time series (P,Q and E)}

\item{infoP}{table containing summary of P events}

\item{hours2extend}{number of hours to extend the rainfall event to account for the discharge peak to occur}
}
\value{
table containing summary of events for the Q time series
}
\description{
Find discharge events
}
\examples{
# tableQ <- findQevents(data$Q,tableP)
}

