% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findQevents.R
\name{findQevents}
\alias{findQevents}
\title{Find discharge events}
\usage{
findQevents(Q, eventTable, hours2extend)
}
\arguments{
\item{Q}{is the discharge time series}

\item{eventTable}{table containing summary of P events}

\item{hours2extend}{number of hours to extend the rainfall event to account for the discharge peak to occur}
}
\value{
table containing summary of events for the Q time series
}
\description{
Find discharge events
}
\examples{
# tableQ <- findQevents(Q,eventTable,hours2extend=6)

}

