% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindQevents.R
\name{FindQevents}
\alias{FindQevents}
\title{Find discharge events}
\usage{
FindQevents(Q, eventTableP, hours2extend, verbose = FALSE)
}
\arguments{
\item{Q}{is the discharge time series}

\item{eventTableP}{table containing summary of P events}

\item{hours2extend}{number of hours to extend the rainfall event to account for the discharge peak to occur}

\item{verbose}{(optional) boolean (FALSE by default). If TRUE prints the progress in terms of event number.}
}
\value{
table containing summary of events for the Q time series
}
\description{
Find discharge events
}
\examples{
\dontrun{
  data("SevernTS")
  tableP <- FindPevents(SevernTS$P[1:100])
  tableQ <- FindQevents(Q = SevernTS$Q[1:100],
                        eventTableP = tableP,
                        hours2extend = 6)
}

}

