% Generated by roxygen2 (4.0.2): do not edit by hand
\name{findQevents}
\alias{findQevents}
\title{Find discharge events}
\usage{
findQevents(dataQ, newInfoP, plotOption = TRUE)
}
\arguments{
\item{dataQ}{is a Q time series}

\item{newInfoP}{table containing summary of P events}

\item{plotOption}{boolean, if TRUE (default) it prints a plot to show the division of the events}
}
\value{
table containing summary of events for the Q time series
}
\description{
Find discharge events
}
\examples{
# tableQ <- findQevents(data$Q,tableP)
}

