% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/HOST2USDA.R
\name{HOST2USDA}
\alias{HOST2USDA}
\title{Map HOST classes into USDA classes}
\usage{
HOST2USDA(vectorHOSTclass, methodBy = "Bulygina et al., 2011")
}
\arguments{
\item{vectorHOSTclass}{this is a vector of HOST classes (e.g. c("3","7")). This can contain a single value.}

\item{methodBy}{this is a string that identifies the methodology to be used. Possible values are: "Bulygina et al., 2011" (default), "Halcrow, 2011".}
}
\value{
USDA class, this can be: A, B, C, D or a combination of them.
}
\description{
Map HOST classes into USDA classes
}
\examples{
HOST2USDA(c("3","7"))
}

