% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotEvents.R
\name{PlotEvents}
\alias{PlotEvents}
\title{Plot PQ events}
\usage{
PlotEvents(dataX, infoP, infoQ)
}
\arguments{
\item{dataX}{is a data.frame containing P, Q (and E) time series}

\item{infoP}{table containing summary of P events}

\item{infoQ}{table containing summary of Q events}
}
\value{
Plot showing PQ events
}
\description{
Plot PQ events
}
\examples{
# PlotEvents(dataX, infoP, infoQ)
}

