% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ReturnPeriod.R
\name{ReturnPeriod}
\alias{ReturnPeriod}
\title{Calculate return periods for PQ events}
\usage{
ReturnPeriod(infoP, infoQ)
}
\arguments{
\item{infoP}{table containing summary of P events}

\item{infoQ}{table containing summary of Q events}
}
\value{
data.frame containing 3 columns: Tr (return period), P (max precipitation) and Q (max discharge)
}
\description{
Calculate return periods for PQ events
}
\examples{
# ReturnPeriod(infoP,infoQ)
}

