% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ReturnPeriod}
\alias{ReturnPeriod}
\title{Calculate return periods for PQ events}
\usage{
ReturnPeriod(infoP, infoQ, plotOption = FALSE, variable2plot = "Q",
  PQindependent = FALSE)
}
\arguments{
\item{infoP}{table containing summary of P events}

\item{infoQ}{table containing summary of Q events}

\item{plotOption}{boolean, if TRUE (default) it prints a plot to show the trend of the return period}

\item{variable2plot}{it can be P or Q (default = Q)}

\item{PQindependent}{boolean. If TRUE the events P and Q are considered independent and the matching return periods are calculated using the frequency matching approach. If FALSE (default) the events P and Q are dependent and they have the same return period}
}
\value{
data.frame containing 3 columns: Tr (return period), P (max precipitation) and Q (max discharge)
}
\description{
Calculate return periods for PQ events
}
\examples{
# ReturnPeriod(infoP,infoQ)
}

