% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilandVeg2CN.R
\name{SoilandVeg2CN}
\alias{SoilandVeg2CN}
\title{Soil and Vegetation code to CN, and corrects it based on drainage level, if needed.}
\usage{
SoilandVeg2CN(
  vegetationCODE,
  soilCODE,
  lookupTable,
  artificialDrainage = "none"
)
}
\arguments{
\item{vegetationCODE}{vegetation class code (e.g. for the UK vegetation map 2013 this is an integer in the range [1,10]).}

\item{soilCODE}{soil class code (e.g. for the HOST system this is an integer in the range [1,29]).}

\item{lookupTable}{this is a data.frame to link Soil, Vegetation and Hydrological conditions.}

\item{artificialDrainage}{(default = "none"), possible values are none, low, medium,high.}
}
\value{
Curve Number for a single raster cell, this is an integer in the range [0,100] with 0 = no runoff and 100 = all rainfall becomes runoff.
}
\description{
Soil and Vegetation code to CN, and corrects it based on drainage level, if needed.
}
\examples{
\dontrun{
# Load lookup table
dfLookup <- MakeLoopkupTable("Severn&Wye")
cnLocal <- SoilandVeg2CN(vegetationCODE = 6,
                         soilCODE = 15,
                         lookupTable = dfLookup,
                         artificialDrainage = "none")
}

}
